joystickOI class
