+++ Team 3749 2023/src/main/java/frc/robot/RobotContainer.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.util.Map;
 
 import edu.wpi.first.wpilibj.DriverStation;
+import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.shuffleboard.BuiltInLayouts;
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
@@ -28,13 +29,13 @@ public class RobotContainer {
     private final Xbox pilot = new Xbox(0);
     private final Xbox operator = new Xbox(1);
 
-    private static String[] lastJoystickNames = new String[] { "", "", "", "", "", "" };
-
     // Subsystems
     private final Swerve swerve = new Swerve();
     private final ArmIntake armIntake = new ArmIntake();
     private final SideIntake sideIntake = new SideIntake();
     private final Arm arm = new Arm();
+
+    JoystickOI joystickOI = new JoystickOI(pilot, operator, swerve, armIntake, sideIntake, arm);
 
     public RobotContainer() {
         DriverStation.silenceJoystickConnectionWarning(true);
@@ -75,117 +76,11 @@ public class RobotContainer {
      */
     public void configureButtonBindings() {
 
-        if (!didJoysticksChange())
+        if (!JoystickOI.didJoysticksChange())
             return;
         CommandScheduler.getInstance().getActiveButtonLoop().clear();
 
-        // if both xbox controllers are connected
-        if (DriverStation.isJoystickConnected(1)) {
-
-            // arm setpoints (buttons)
-            operator.a().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_TOP));
-            operator.b().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_MID));
-            operator.x().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.GROUND_INTAKE));
-
-            // arm setpoints (bumpers)
-            operator.rightBumper().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.STING));
-            operator.leftBumper().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.DOUBLE_SUBSTATION));
-
-            operator.rightTriggerWhileHeld(() -> armIntake.setVoltage(Constants.ArmIntake.releaseObjectVoltage));
-            operator.leftTriggerWhileHeld(() -> armIntake.setVoltage(Constants.ArmIntake.intakeVoltage));
-
-            pilot.y().onTrue(Commands.runOnce(() -> sideIntake.toggleLiftSetpoint(), sideIntake));
-
-            pilot.rightTriggerWhileHeld(() -> sideIntake.setIntakeVoltage(Constants.ArmIntake.releaseObjectVoltage),
-                    sideIntake);
-            pilot.leftTriggerWhileHeld(() -> sideIntake.setIntakeVoltage(Constants.ArmIntake.intakeVoltage),
-                    sideIntake);
-
-            // swerve button bindings
-            pilot.backWhileHeld(() -> swerve.zeroHeading(), swerve);
-
-            // swerve rotation cardinals
-            pilot.povUp().whileTrue(Commands.run(() -> swerve.turnToRotation(0)));
-            pilot.povLeft().whileTrue(Commands.run(() -> swerve.turnToRotation(270)));
-            pilot.povDown().whileTrue(Commands.run(() -> swerve.turnToRotation(180)));
-            pilot.povRight().whileTrue(Commands.run(() -> swerve.turnToRotation(90)));
-
-            // if only one xbox controller is connected
-        } else if (DriverStation.isJoystickConnected(0)) {
-
-            // arm setpoints (buttons)
-            pilot.a().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_TOP));
-            pilot.b().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_MID));
-            pilot.x().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.GROUND_INTAKE));
-            pilot.y().onTrue(Commands.runOnce(() -> sideIntake.toggleLiftSetpoint(), sideIntake));
-
-            // arm setpoints (bumpers)
-            pilot.rightBumper().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.STING));
-            pilot.leftBumper().onTrue(new MoveArm(arm, armIntake, ArmSetpoints.DOUBLE_SUBSTATION));
-
-            // intake button bindings
-            pilot.rightTriggerWhileHeld(() -> armIntake.setVoltage(Constants.ArmIntake.releaseObjectVoltage));
-            pilot.leftTriggerWhileHeld(() -> armIntake.setVoltage(Constants.ArmIntake.intakeVoltage));
-
-            // swerve button bindings
-            pilot.backWhileHeld(() -> swerve.zeroHeading(), swerve);
-
-            // swerve rotation cardinals
-            pilot.povUp().whileTrue(Commands.run(() -> swerve.turnToRotation(0)));
-            pilot.povLeft().whileTrue(Commands.run(() -> swerve.turnToRotation(270)));
-            pilot.povDown().whileTrue(Commands.run(() -> swerve.turnToRotation(180)));
-            pilot.povRight().whileTrue(Commands.run(() -> swerve.turnToRotation(90)));
-
-            // if no joysticks are connected (ShuffleBoard buttons)
-        } else {
-
-            ShuffleboardTab controlsTab = Shuffleboard.getTab("Controls");
-
-            ShuffleboardLayout armCommands = controlsTab
-                    .getLayout("Arm", BuiltInLayouts.kList)
-                    .withSize(2, 2)
-                    .withProperties(Map.of("Label position", "HIDDEN")); // hide labels for commands
-
-            armCommands.add(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_TOP));
-            armCommands.add(new MoveArm(arm, armIntake, ArmSetpoints.PLACE_MID));
-            armCommands.add(new MoveArm(arm, armIntake, ArmSetpoints.GROUND_INTAKE));
-            armCommands.add(new MoveArm(arm, armIntake, ArmSetpoints.STING));
-            armCommands.add(new MoveArm(arm, armIntake, ArmSetpoints.DOUBLE_SUBSTATION));
-
-            ShuffleboardLayout sideIntakeCommands = controlsTab
-                    .getLayout("Side Intake", BuiltInLayouts.kList)
-                    .withSize(2, 2)
-                    .withProperties(Map.of("Label position", "HIDDEN"));
-
-            CommandBase liftSideIntake = Commands.runOnce(() -> sideIntake.toggleLiftSetpoint(), sideIntake);
-            liftSideIntake.setName("Toggle Lift");
-            CommandBase outakeSideIntake = Commands
-                    .run(() -> sideIntake.setIntakeVoltage(Constants.ArmIntake.releaseObjectVoltage), sideIntake);
-            outakeSideIntake.setName("Side Outake");
-            CommandBase intakeSideIntake = Commands
-                    .run(() -> sideIntake.setIntakeVoltage(Constants.ArmIntake.intakeVoltage), sideIntake);
-            intakeSideIntake.setName("Side Intake");
-
-            sideIntakeCommands.add(liftSideIntake);
-            sideIntakeCommands.add(outakeSideIntake);
-            sideIntakeCommands.add(intakeSideIntake);
-
-            ShuffleboardLayout armIntakeCommands = controlsTab
-                    .getLayout("arm", BuiltInLayouts.kList)
-                    .withSize(2, 2)
-                    .withProperties(Map.of("Label position", "HIDDEN"));
-
-            CommandBase outakeArmIntake = Commands
-                    .run(() -> armIntake.setVoltage(Constants.ArmIntake.releaseObjectVoltage), sideIntake);
-            outakeArmIntake.setName("Arm Outake");
-            CommandBase intakeArmIntake = Commands.run(() -> armIntake.setVoltage(Constants.ArmIntake.intakeVoltage),
-                    sideIntake);
-            intakeArmIntake.setName("Arm Intake");
-
-            armIntakeCommands.add(outakeArmIntake);
-            armIntakeCommands.add(intakeArmIntake);
-
-        }
+        joystickOI.getButtonBindings();
 
     }
 
@@ -213,17 +108,4 @@ public class RobotContainer {
         Constants.AutoConstants.eventMap.put("place_cone_top", null);
         Constants.AutoConstants.eventMap.put("place_cube_top", null);
     }
-
-    public static boolean didJoysticksChange() {
-        boolean joysticksChanged = false;
-        for (int port = 0; port < DriverStation.kJoystickPorts; port++) {
-            String name = DriverStation.getJoystickName(port);
-            if (!name.equals(lastJoystickNames[port])) {
-                joysticksChanged = true;
-                lastJoystickNames[port] = name;
-            }
-        }
-        return joysticksChanged;
-    }
-
 }
